streamlit/requirements.txt
absl-py==2.1.0
alembic==1.13.1
altair==5.3.0
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
backports.weakref==1.0.post1
blinker==1.8.2
brotlipy==0.7.0
cachetools==5.3.3
catboost==1.2.3
category-encoders==2.6.3
colorlog==6.8.2
constantly==15.1.0
cssselect==1.1.0
daal4py==2023.1.1
et-xmlfile==1.1.0
fake-useragent==1.5.0
flatbuffers==24.3.25
fonttools==4.25.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
imbalanced-learn==0.12.3
imblearn==0.0
inflection==0.5.1
interact==0.2
joblib==1.4.2
jsonpointer==2.1
keras==3.3.3
libclang==18.1.1
lightgbm==4.3.0
llvmlite==0.42.0
Mako==1.3.3
matplotlib==3.9.0
munkres==1.1.4
namex==0.0.8
numba==0.59.1
numpy==1.24.2
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
optuna-integration==3.6.0
outcome==1.3.0.post0
pandas-compat==0.1.1
patsy==0.5.6
pep8==1.7.1
phik==0.12.4
ply==3.11
protobuf==4.25.3
pyarrow==11.0.0
pyasn1-modules==0.2.8
pydeck==0.9.1
PyDispatcher==2.0.5
pyee==8.2.2
pygame==2.5.2
pyls-spyder==0.4.0
pyppeteer==1.0.2
PyQt5==5.15.7
python-lsp-jsonrpc==1.0.0
pywin32==305.1
queuelib==1.5.0
rich==13.7.1
scikit-learn==1.5.0
scipy==1.11.1
seaborn==0.13.2
selenium==4.18.1
shap==0.45.1
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
statsmodels==0.14.2
streamlit==1.35.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
zstandard==0.19.0
